+++ src/bun/tests/update.test.ts
@@ -1,6 +1,7 @@
-import { describe, expect, it } from 'vitest'
-import sql from 'sql.js'
+import { describe, expect, it } from 'bun:test'
+import { Database } from 'bun:sqlite'
 import Sibyl from '../index'
+import type { SibylResponse } from '../../types'
 
 interface TableRow {
   id: number
@@ -15,12 +16,7 @@ interface Tables {
 
 describe('update tests', () => {
   it('updates an entry in the DB', async () => {
-    const SQL = await sql({
-      locateFile: () => {
-        return 'playground/public/sql-wasm.wasm'
-      },
-    })
-    const db = new SQL.Database()
+    const db = new Database(':memory:')
     const { createTable, Insert, Update } = await Sibyl<Tables>(db)
 
     createTable('first', {
@@ -64,11 +60,11 @@ describe('update tests', () => {
         booleanTest: false,
       },
     })
-    const expectation = {
+    const expectation: SibylResponse<TableRow> = {
+      booleanTest: 0,
       id: 2344,
       location: 'Brighton',
       name: 'Richard',
-      booleanTest: 0,
     }
     expect(actual).toStrictEqual(expectation)
   })