+++ src/types.ts
@@ -1,3 +1,9 @@
+export type SibylResponse<T> = {
+  [Key in keyof T]:
+  T[Key] extends boolean ? 0 | 1 :
+    T[Key]
+}
+
 export interface DBEntry<T> {
   type: T
   primary?: boolean
@@ -6,20 +12,24 @@ export interface DBEntry<T> {
   autoincrement: boolean
 }
 
+export type DBValue<T> = T extends DBNumber
+  ? DBEntry<T>
+  : Omit<DBEntry<T>, 'autoincrement'>
+
 export type DBBoolean = 'bool'
 export type DBNumber = 'int' | 'real' | 'INTEGER'
 export type DBString = 'varchar' | 'char'
 export type DBDate = 'text' | 'int' | 'real'
 export type DBBlob = 'blob'
 
-export type DBValue<T> = T extends DBNumber
-  ? DBEntry<T>
-  : Omit<DBEntry<T>, 'autoincrement'>
-
-export type SibylResponse<T> = {
+export type MappedTable<T> = {
   [Key in keyof T]:
-  T[Key] extends boolean ? 0 | 1 :
-    T[Key]
+  T[Key] extends boolean ? DBValue<DBBoolean> :
+    T[Key] extends number ? DBValue<DBNumber> :
+      T[Key] extends string ? DBValue<DBString> :
+        T[Key] extends Date ? DBValue<DBDate> :
+          T[Key] extends Blob ? DBValue<DBBlob> :
+            null
 }
 
 interface OR<T> {